Move svg utils
